Conflicts:
	package.json
	test/ical_test.coffee

+++ client/app/views/list_view.coffee
@@ -68,17 +68,19 @@ module.exports = class ListView extends View
             @alarmFormView.displayErrors(alarm.validationError)
 
     onEditAlarmClicked: (event) ->
+        window.top.window.scrollTo(0,0)
         alarmID = $(event.target).data('alarmid')
         alarm = @model.get alarmID
         @alarmFormView.loadAlarmData(alarm)
 
-    onRemoveAlarmClicked: (event) ->
-        alarmID = $(event.target).data('alarmid')
-        alarm = @model.get alarmID
-        # TODO: add confirmation and loading indicator
-        alarm.destroy
-            wait: true
-            success: () ->
-                console.log "Delete alarm: success"
-            error: () ->
-                console.log "Delete alarm: error"
+    onRemoveAlarmClicked: (event) ->        
+        if confirm 'Are you sure ?'
+            alarmID = $(event.target).data('alarmid')
+            alarm = @model.get alarmID
+            # TODO: add confirmation and loading indicator
+            alarm.destroy
+                wait: true
+                success: () ->
+                    console.log "Delete alarm: success"
+                error: () ->
+                    console.log "Delete alarm: error"
